+++ src/components/admin/AdminTypes.ts
@@ -6,6 +6,7 @@ export type ISystemConfig = {
   'memories.exiftool_no_local': boolean;
   'memories.index.mode': string;
   'memories.index.path': string;
+  'memories.index.path.blacklist': string;
 
   'memories.gis_type': number;